pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vocalyx"
version = "1.0.1"
description = "Real-time voice assistant with Speech-to-Text (STT), GPT, and Text-to-Speech (TTS). Run vocalyx-stt, vocalyx-tts, and vocalyx in three terminals."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Daxrajsinh Jadeja"}
]
keywords = ["voice", "assistant", "stt", "tts", "speech-to-text", "text-to-speech", "realtime", "soprano"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "RealtimeSTT>=0.3.0",
    "soprano-tts>=0.2.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
    "PyAudio>=0.2.14",
    "colorama>=0.4.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "build>=1.0.0",
    "twine>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/Daxrajsinh-holboxai/Vocalyx"
Repository = "https://github.com/Daxrajsinh-holboxai/Vocalyx.git"

[project.scripts]
vocalyx-tts = "vocalyx.tts_server:cli"
vocalyx-stt = "vocalyx.stt_server:cli"
vocalyx = "vocalyx.client:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["vocalyx*"]
